requirements.txt
asgiref==3.5.2
Django==4.1.4
django-filter==22.1
django-phonenumber-field==7.0.1
django-phonenumbers==1.0.1
djangorestframework==3.14.0
importlib-metadata==5.1.0
Markdown==3.4.1
phonenumbers==8.13.2
Pillow==9.3.0
pytz==2022.6
sqlparse==0.4.3
zipp==3.11.0
